Forgot to add these here.

+++ src/arcemu-world/Messenger.cpp
@@ -827,10 +827,10 @@ void Messenger::SendEnchantmentLog(Player *player, uint32 itemId, uint32 enchant
 void Messenger::SendEnchantTimeUpdate( Player *player, uint64 itemGUID, uint32 slot, uint32 duration )
 {
 	WorldPacket* data = new WorldPacket( SMSG_ITEM_ENCHANT_TIME_UPDATE, 24 );
-	*data << itemGUID;
-	*data << slot;
-	*data << duration;
-	*data << player->GetGUID();
+	*data << uint64( itemGUID );
+	*data << uint32( slot );
+	*data << uint32( duration );
+	*data << uint64( player->GetGUID() );
 	PlayerMessenger::sendDelayedMessage( player, data );
 }
